requirements.txt
prisma 
fastapi
uvicorn
python-dotenv
bcrypt
PyJWT
pydantic
python-multipart
redis